WIP, do not push. fixing keyboard navigation for autocomplete
I'm halfway through moving the keyboard navigation out of the <Menu> component and into the <Autocomplete>. The navigation works great in Docz, but doesn't work in Jobber for whatever reason. It seems that putting handlers on the global scope is probably the culprit.

In order to fix the issue we need to isolate the keyboard event handlers to the <Autocomplete> and we need to have a way to have events in the textbox make things happen in the menu. Putting handlers on the surrounding <div> should work here. The up/down/enter/esc actions can be caught at the <div> level and do some state changes to move the highlight on the menu, etc.

Also added some test stubs for the component.